+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## WIP
+
+  * Fix PHP notice triggered when an observer is disabled
+
 ## [4.0.2] - 2019-02-06
 [4.0.2]: https://github.com/Smile-SA/magento2-module-debug-toolbar/compare/4.0.1...4.0.2